--- js/kk.js
@@ -1,4 +1,0 @@
-
-export function mbclick(){
-    return "dd";
-}
