--- pug/header.pug
@@ -1,10 +1,0 @@
-
-
-span.headers
-    span.carmoa
-    span.hom 홈
-    span.dangi 단기렌트
-    span.wal 월렌트
-    span.event 이벤트
-
-    span.mypage 마이페이지
